[test] fold EuRoC preparation tests into the conversion suite
Bundle selection lives in convert_euroc, so a separate preparation test file
drew a boundary that did not match the code and duplicated coverage the
conversion tests already had: the all-bundles default and the unknown-sequence
error were each asserted twice.

Keep the per-group mapping, the bundle ordering, and the resulting download
arguments, plus the corrupt-cache rejection. Drop the user-agent and
cached-bundle tests, which only passed by stubbing curl and md5sum on PATH.
TUM keeps one case per validation branch: member name and link target.
