+++ tools/python_tools/cuvslam_tools/tests/test_euroc_conversion.py
@@ -16,6 +16,7 @@ import contextlib
 import io
 import json
 import os
+import subprocess
 import tempfile
 import unittest
 import zipfile
@@ -521,6 +522,67 @@ class TestEurocConversionFailures(unittest.TestCase):
                 )
                 with self.assertRaisesRegex(convert_euroc.ConversionError, "unsafe image filename"):
                     convert_euroc.convert(raw, self.output / str(index), ["MH_01_easy"])
+
+
+class TestEurocBundleSelection(unittest.TestCase):
+    """Selecting the wrong bundle downloads 6-12 GB that is not needed.
+
+    The default and unknown-sequence cases are covered by the conversion tests
+    above; these pin the per-group mapping and the resulting download arguments.
+    """
+
+    def test_each_sequence_group_maps_to_its_own_bundle(self):
+        cases = {
+            "MH_03_medium": ["machine_hall.zip"],
+            "V1_02_medium": ["vicon_room1.zip"],
+            "V2_03_difficult": ["vicon_room2.zip"],
+        }
+        for sequence, expected in cases.items():
+            with self.subTest(sequence=sequence):
+                self.assertEqual(convert_euroc.required_archives([sequence]), expected)
+
+    def test_mixed_subset_keeps_bundle_order(self):
+        self.assertEqual(
+            convert_euroc.required_archives(["V2_01_easy", "MH_01_easy"]),
+            ["machine_hall.zip", "vicon_room2.zip"],
+        )
+
+    def test_explicit_subset_downloads_only_its_bundles(self):
+        with tempfile.TemporaryDirectory() as temporary:
+            raw_dir = Path(temporary) / "raw"
+            with mock.patch.object(euroc_prepare, "run_download_script") as download:
+                with contextlib.redirect_stdout(io.StringIO()):
+                    euroc_prepare.prepare(
+                        raw_dir=raw_dir,
+                        output_dir=Path(temporary) / "converted",
+                        sequences=["MH_01_easy", "V1_01_easy"],
+                        download_only=True,
+                    )
+
+        self.assertEqual(
+            download.call_args.args[1],
+            [str(raw_dir), "--archive", "machine_hall.zip", "--archive", "vicon_room1.zip"],
+        )
+
+
+class TestEurocDownloadScript(unittest.TestCase):
+    def test_corrupt_cached_archive_is_rejected(self):
+        """A truncated multi-GB download must not be handed to the converter."""
+        script = Path(convert_euroc.__file__).with_name("download_euroc.sh")
+        with tempfile.TemporaryDirectory() as temporary:
+            raw_dir = Path(temporary)
+            (raw_dir / "machine_hall.zip").write_bytes(b"corrupt")
+
+            completed = subprocess.run(
+                ["bash", str(script), "--archive", "machine_hall.zip", str(raw_dir)],
+                check=False,
+                capture_output=True,
+                text=True,
+            )
+
+        self.assertEqual(completed.returncode, 1)
+        self.assertIn("md5 mismatch for existing", completed.stderr)
+        self.assertIn("re-run with --force", completed.stderr)
 
 
 if __name__ == "__main__":
